requirements.txt
streamlit
fpdf
pdfplumber
pytesseract
pdf2image
Pillow
nltk
scikit-learn
langchain
langchain-groq
groq
pydantic
numpy
